@@ -1,4 +1,3 @@
-from typing import Optional
 from unittest import mock
 
 import numpy as np
@@ -66,7 +65,7 @@ def default_chronos_tiny_model(request, chronos_model_path) -> ChronosModel:
 
 
 @pytest.fixture(scope="module", params=HYPERPARAMETER_DICTS)
-def default_chronos_tiny_model_gpu(request, chronos_model_path) -> Optional[ChronosModel]:
+def default_chronos_tiny_model_gpu(request, chronos_model_path) -> ChronosModel | None:
     if not GPU_AVAILABLE:
         pytest.skip(reason="GPU not available")
 
@@ -503,3 +502,18 @@ def test_when_chronos_bolt_no_fine_tune_with_custom_quantiles_then_original_quan
     )
     model.fit(train_data=DUMMY_TS_DATAFRAME)
     assert model.model_pipeline.quantiles == original_quantiles
+
+
+def test_when_revision_provided_then_from_pretrained_is_called_with_revision(chronos_model_path):
+    model_revision = "my-test-branch"
+    model = ChronosModel(
+        hyperparameters={"model_path": chronos_model_path, "revision": model_revision, "device": "cpu"},
+    )
+
+    with mock.patch("chronos.BaseChronosPipeline.from_pretrained") as mock_from_pretrained:
+        mock_from_pretrained.return_value = mock.MagicMock()
+        model.fit(train_data=DUMMY_TS_DATAFRAME)
+        model.load_model_pipeline()
+
+    mock_from_pretrained.assert_called_once()
+    assert mock_from_pretrained.call_args.kwargs.get("revision") == model_revision
