@@ -130,7 +130,7 @@ class HFAutoModelForTextPrediction(nn.Module):
 
         self.gradient_checkpointing = gradient_checkpointing
         if gradient_checkpointing:
-            self.model.gradient_checkpointing_enable()
+            self.model.gradient_checkpointing_enable(gradient_checkpointing_kwargs={"use_reentrant": False})
             if self.is_t5:
                 self.dummy_layer = DummyLayer()
 
