@@ -1,3 +1,4 @@
+from .chronos2 import Chronos2Model
 from .model import ChronosModel
 
-__all__ = ["ChronosModel"]
+__all__ = ["ChronosModel", "Chronos2Model"]
