@@ -214,6 +214,9 @@ class TestAllModelsWhenHyperparameterTuning:
             assert 1 <= result["hyperparameters"]["max_epochs"] <= 3
 
     @pytest.mark.parametrize("searcher", ["random", "bayes"])
+    @pytest.mark.skipif(
+        sys.platform == "win32" and sys.version_info >= (3, 13), reason="No ray support on Windows with Python 3.13"
+    )
     def test_given_searcher_when_ray_backend_used_in_hpo_then_correct_searcher_used(
         self, gluonts_model_class, searcher
     ):
@@ -477,9 +480,10 @@ class TestAllModelsWhenPreprocessingAndTransformsRequested:
         assert isinstance(regressor, CovariateRegressor)
         assert regressor.is_fit()
 
+        known_covariates = test_data.slice_by_timestep(-prediction_length, None).drop(columns=["target"])
         predictions = model.predict(
             train_data,
-            known_covariates=test_data.slice_by_timestep(-prediction_length, None),
+            known_covariates=known_covariates,
         )
         assert isinstance(predictions, TimeSeriesDataFrame)
         assert not predictions.isna().any(axis=None)
@@ -502,4 +506,4 @@ class TestInferenceOnlyModels:
                 mw_model.fit(train_data=data, time_limit=time_limit)
             except RuntimeError:
                 pass
-            assert abs(mock_predict.call_args[1]["time_limit"] - time_limit) < 0.5
+            assert abs(mock_predict.call_args[1]["time_limit"] - time_limit) < 5.0
