@@ -1,126 +1,176 @@
+import sys
 import time
 
 import pytest
 
-from autogluon.timeseries.utils.timer import SplitTimer
+from autogluon.timeseries.utils.timer import SplitTimer, Timer
 
 
-@pytest.mark.parametrize("strict_stop", [True, False])
-def test_when_timer_not_started_then_get_raises_error(strict_stop):
-    timer = SplitTimer(10.0, rounds=2, strict_stop=strict_stop)
-    with pytest.raises(RuntimeError, match="Timer has not been started"):
-        timer.get()
+@pytest.mark.skipif(sys.platform in ["darwin", "win32"], reason="time module is unreliable on MacOS/Windows")
+class TestTimer:
+    def test_when_timer_not_started_then_time_elapsed_raises_error(self):
+        timer = Timer(time_limit=10.0)
+        with pytest.raises(RuntimeError, match="Timer has not been started"):
+            timer.time_elapsed()
+
+    def test_when_timer_not_started_then_time_remaining_raises_error(self):
+        timer = Timer(time_limit=10.0)
+        with pytest.raises(RuntimeError, match="Timer has not been started"):
+            timer.time_remaining()
+
+    def test_when_timer_not_started_then_timed_out_raises_error(self):
+        timer = Timer(time_limit=10.0)
+        with pytest.raises(RuntimeError, match="Timer has not been started"):
+            timer.timed_out()
+
+    def test_when_timer_started_then_time_elapsed_returns_positive(self):
+        timer = Timer(time_limit=10.0).start()
+        time.sleep(0.01)
+        assert timer.time_elapsed() >= 0.01
+
+    def test_when_timer_started_then_time_remaining_decreases(self):
+        timer = Timer(time_limit=1.0).start()
+        initial_remaining = timer.time_remaining()
+        time.sleep(0.01)
+        later_remaining = timer.time_remaining()
+        assert later_remaining < initial_remaining  # type: ignore
+
+    def test_when_time_limit_none_then_time_remaining_returns_none(self):
+        timer = Timer(time_limit=None).start()
+        assert timer.time_remaining() is None
+
+    def test_when_time_limit_none_then_timed_out_returns_false(self):
+        timer = Timer(time_limit=None).start()
+        time.sleep(0.01)
+        assert timer.timed_out() is False
+
+    def test_when_time_not_exceeded_then_timed_out_returns_false(self):
+        timer = Timer(time_limit=10.0).start()
+        assert timer.timed_out() is False
+
+    def test_when_time_exceeded_then_timed_out_returns_true(self):
+        timer = Timer(time_limit=0.01).start()
+        time.sleep(0.02)
+        assert timer.timed_out() is True
+
+    def test_when_start_called_twice_then_timer_resets(self):
+        timer = Timer(time_limit=10.0).start()
+        time.sleep(0.01)
+        first_elapsed = timer.time_elapsed()
+
+        timer.start()
+        second_elapsed = timer.time_elapsed()
+        assert second_elapsed < first_elapsed
 
 
-@pytest.mark.parametrize("strict_stop", [True, False])
-def test_when_timer_not_started_then_time_elapsed_raises_error(strict_stop):
-    timer = SplitTimer(10.0, rounds=2, strict_stop=strict_stop)
-    with pytest.raises(RuntimeError, match="Timer has not been started"):
-        timer.time_elapsed()
+@pytest.mark.skipif(sys.platform in ["darwin", "win32"], reason="time module is unreliable on MacOS/Windows")
+class TestSplitTimer:
+    def test_when_timer_not_started_then_round_time_remaining_raises_error(self):
+        timer = SplitTimer(10.0, rounds=2)
+        with pytest.raises(RuntimeError, match="Timer has not been started"):
+            timer.round_time_remaining()
 
+    def test_when_timer_not_started_then_time_elapsed_raises_error(self):
+        timer = SplitTimer(10.0, rounds=2)
+        with pytest.raises(RuntimeError, match="Timer has not been started"):
+            timer.time_elapsed()
 
-@pytest.mark.parametrize("strict_stop", [True, False])
-def test_when_time_limit_is_none_then_get_returns_none(strict_stop):
-    timer = SplitTimer(None, rounds=2, strict_stop=strict_stop).start()
-    assert timer.get() is None
+    def test_when_timer_not_started_then_round_time_elapsed_raises_error(self):
+        timer = SplitTimer(10.0, rounds=2)
+        with pytest.raises(RuntimeError, match="Timer has not been started"):
+            timer.round_time_elapsed()
 
+    def test_when_timer_not_started_then_next_round_raises_error(self):
+        timer = SplitTimer(10.0, rounds=2)
+        with pytest.raises(RuntimeError, match="Timer has not been started"):
+            timer.next_round()
 
-@pytest.mark.parametrize("strict_stop", [True, False])
-def test_when_timer_started_then_get_returns_split_time(strict_stop):
-    timer = SplitTimer(10.0, rounds=2, strict_stop=strict_stop).start()
-    first_split = timer.get()
+    def test_when_time_limit_is_none_then_round_time_remaining_returns_none(self):
+        timer = SplitTimer(None, rounds=2).start()
+        assert timer.round_time_remaining() is None
 
-    assert first_split is not None
-    assert abs(first_split - 5.0) < 0.01
+    def test_when_timer_started_then_round_time_remaining_returns_split_time(self):
+        timer = SplitTimer(10.0, rounds=2).start()
+        first_split = timer.round_time_remaining()
 
+        assert first_split is not None
+        assert abs(first_split - 5.0) < 0.01
 
-@pytest.mark.parametrize("strict_stop", [True, False])
-def test_when_timer_split_then_remaining_time_adjusts(strict_stop):
-    timer = SplitTimer(10.0, rounds=2, strict_stop=strict_stop).start()
-    timer.split()
-    second_split = timer.get()
-    assert second_split is not None
-    assert abs(second_split - 10.0) < 0.01  # All remaining time for last round
+    def test_when_timer_next_round_then_remaining_time_adjusts(self):
+        timer = SplitTimer(10.0, rounds=2).start()
+        timer.next_round()
+        second_split = timer.round_time_remaining()
+        assert second_split is not None
+        assert abs(second_split - 10.0) < 0.01  # All remaining time for last round
 
+    def test_when_all_rounds_used_then_round_time_remaining_returns_zero(self):
+        timer = SplitTimer(10.0, rounds=2).start()
+        timer.next_round()
+        timer.next_round()
+        assert timer.round_time_remaining() == 0.0
 
-def test_when_all_rounds_used_then_get_returns_zero():
-    timer = SplitTimer(10.0, rounds=2).start()
-    timer.split()
-    timer.split()
-    assert timer.get() == 0.0
+    def test_when_time_elapsed_then_returns_correct_duration(self):
+        timer = SplitTimer(10.0, rounds=2).start()
+        time.sleep(0.01)
+        elapsed = timer.time_elapsed()
+        assert elapsed >= 0.01
 
+    def test_when_round_time_elapsed_then_returns_correct_duration(self):
+        timer = SplitTimer(10.0, rounds=2).start()
+        time.sleep(0.01)
+        elapsed = timer.round_time_elapsed()
+        assert elapsed >= 0.01
 
-@pytest.mark.parametrize("strict_stop", [True, False])
-def test_when_time_elapsed_then_returns_correct_duration(strict_stop):
-    timer = SplitTimer(10.0, rounds=2, strict_stop=strict_stop).start()
-    time.sleep(0.01)
-    elapsed = timer.time_elapsed()
-    assert elapsed >= 0.01
+    def test_when_next_round_then_round_time_elapsed_resets(self):
+        timer = SplitTimer(10.0, rounds=2).start()
+        time.sleep(0.02)
+        first_round_elapsed = timer.round_time_elapsed()
+        assert first_round_elapsed >= 0.02
 
+        timer.next_round()
+        second_round_elapsed = timer.round_time_elapsed()
+        assert second_round_elapsed < first_round_elapsed
 
-@pytest.mark.parametrize("rounds", [1, 3, 5])
-def test_when_iterator_used_then_yields_correct_number_of_splits(rounds):
-    timer = SplitTimer(10.0, rounds=rounds)
-    splits = list(timer)
-    assert len(splits) == rounds
+    def test_when_start_called_then_timer_resets(self):
+        timer = SplitTimer(10.0, rounds=2).start()
+        timer.next_round()
+        assert timer.round_index == 1
 
+        timer.start()
+        assert timer.round_index == 0
+        assert abs(timer.round_time_remaining() - 5.0) < 0.01  # type: ignore
 
-def test_when_iterator_used_then_splits_sum_to_total_time():
-    timer = SplitTimer(10.0, rounds=4)
-    splits = list(timer)
-    expected_splits = [2.5, 3.33333333, 5.0, 10.0]
-    for actual, expected in zip(splits, expected_splits):
-        assert abs(actual - expected) < 0.01  # type: ignore
+    def test_when_time_exceeded_then_returns_negative(self):
+        timer = SplitTimer(0.1, rounds=2).start()
+        time.sleep(0.15)
+        result = timer.round_time_remaining()
+        assert result < 0  # type: ignore
 
+    def test_when_round_uses_less_time_then_next_round_gets_more(self):
+        """Test that unused time from one round is redistributed to remaining rounds."""
+        timer = SplitTimer(10.0, rounds=3).start()
 
-def test_when_start_called_then_timer_resets():
-    timer = SplitTimer(10.0, rounds=2).start()
-    timer.split()
-    assert timer.round_index == 1
+        # Round 1: should get ~3.33s (10.0 / 3)
+        round_1_time = timer.round_time_remaining()
+        assert round_1_time is not None
+        assert abs(round_1_time - 3.33) < 0.1
 
-    timer.start()
-    assert timer.round_index == 0
-    assert abs(timer.get() - 5.0) < 0.01  # type: ignore
+        time.sleep(0.01)
+        timer.next_round()
 
+        round_2_time = timer.round_time_remaining()
+        assert round_2_time is not None
+        assert round_2_time > 4.9
 
-def test_when_strict_stop_enabled_and_rounds_exceeded_then_raises_error():
-    timer = SplitTimer(0.5, rounds=2, strict_stop=True).start()
-    timer.split()
-    timer.split()
-    with pytest.raises(RuntimeError, match="Timer has run out of rounds"):
-        timer.get()
+        time.sleep(0.02)
+        timer.next_round()
 
+        round_3_time = timer.round_time_remaining()
+        assert round_3_time is not None
+        assert round_3_time > 9.9
 
-def test_when_strict_stop_enabled_and_time_exceeded_then_raises_error():
-    timer = SplitTimer(0.1, rounds=5, strict_stop=True).start()
-    time.sleep(0.15)
-    with pytest.raises(RuntimeError, match="Time limit exceeded"):
-        timer.get()
-
-
-def test_when_strict_stop_disabled_and_time_exceeded_then_returns_negative():
-    timer = SplitTimer(0.1, rounds=2, strict_stop=False).start()
-    time.sleep(0.15)
-    result = timer.get()
-    assert result < 0  # type: ignore
-
-
-def test_when_strict_stop_enabled_in_iterator_and_time_exceeded_then_stops_gracefully():
-    timer = SplitTimer(0.2, rounds=10, strict_stop=True)
-    splits = []
-    for split_time in timer:
-        splits.append(split_time)
-        time.sleep(0.05)
-
-    assert len(splits) < 10  # Should stop before all rounds
-
-
-def test_when_strict_stop_disabled_in_iterator_then_completes_all_rounds():
-    timer = SplitTimer(0.1, rounds=5, strict_stop=False)
-    splits = []
-    for split_time in timer:
-        splits.append(split_time)
-        time.sleep(0.05)
-
-    assert len(splits) == 5  # all rounds have run
-    assert splits[-1] < 0
+    def test_when_single_round_then_gets_all_time(self):
+        timer = SplitTimer(10.0, rounds=1).start()
+        round_time = timer.round_time_remaining()
+        assert round_time is not None
+        assert abs(round_time - 10.0) < 0.01
